+++ tests/config-reloader/test.sh
@@ -14,7 +14,6 @@ trap cleanup EXIT
 echo "==> Starting docker-salt-master (${PLATFORM}) ..."
 start_container_and_wait \
   --env SALT_RESTART_MASTER_ON_CONFIG_CHANGE=True \
-  --volume "${SCRIPT_PATH}/config":/home/salt/data/config:ro \
 || error "container started"
 ok "container started"